requirements.txt
flask==2.3.3
yt-dlp==2025.7.21
openai==1.99.1
google-generativeai==0.3.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0
dataclasses-json==0.6.1